# Conflicts:
#	nixos/modules/services/networking/strongswan.nix

+++ nixos/modules/services/networking/zerotierone.nix
@@ -1,20 +1,27 @@
-{ config, lib, pkgs, ... }:
+{
+  config,
+  lib,
+  pkgs,
+  ...
+}:
 
 with lib;
 
 let
   cfg = config.services.zerotierone;
-  localConfFile = pkgs.writeText "zt-local.conf" (builtins.toJSON cfg.localConf);
+
+  settingsFormat = pkgs.formats.json { };
+  localConfFile = settingsFormat.generate "zt-local.conf" cfg.localConf;
   localConfFilePath = "/var/lib/zerotier-one/local.conf";
 in
 {
-  options.services.zerotierone.enable = mkEnableOption (lib.mdDoc "ZeroTierOne");
+  options.services.zerotierone.enable = mkEnableOption "ZeroTierOne";
 
   options.services.zerotierone.joinNetworks = mkOption {
-    default = [];
+    default = [ ];
     example = [ "a8a2c3c10c1a68de" ];
     type = types.listOf types.str;
-    description = lib.mdDoc ''
+    description = ''
       List of ZeroTier Network IDs to join on startup.
       Note that networks are only ever joined, but not automatically left after removing them from the list.
       To remove networks, use the ZeroTier CLI: `zerotier-cli leave <network-id>`
@@ -24,7 +31,7 @@ in
   options.services.zerotierone.port = mkOption {
     default = 9993;
     type = types.port;
-    description = lib.mdDoc ''
+    description = ''
       Network port used by ZeroTier.
     '';
   };
@@ -32,16 +39,16 @@ in
   options.services.zerotierone.package = mkPackageOption pkgs "zerotierone" { };
 
   options.services.zerotierone.localConf = mkOption {
-    default = null;
-    description = mdDoc ''
+    default = { };
+    description = ''
       Optional configuration to be written to the Zerotier JSON-based local.conf.
       If set, the configuration will be symlinked to `/var/lib/zerotier-one/local.conf` at build time.
-      To understand the configuration format, refer to https://docs.zerotier.com/config/#local-configuration-options.
+      To understand the configuration format, refer to <https://docs.zerotier.com/config/#local-configuration-options>.
     '';
     example = {
       settings.allowTcpFallbackRelay = false;
     };
-    type = types.nullOr types.attrs;
+    type = settingsFormat.type;
   };
 
   config = mkIf cfg.enable {
@@ -54,22 +61,27 @@ in
 
       path = [ cfg.package ];
 
-      preStart = ''
-        mkdir -p /var/lib/zerotier-one/networks.d
-        chmod 700 /var/lib/zerotier-one
-        chown -R root:root /var/lib/zerotier-one
-      '' + (concatMapStrings (netId: ''
-        touch "/var/lib/zerotier-one/networks.d/${netId}.conf"
-      '') cfg.joinNetworks) + optionalString (cfg.localConf != null) ''
-        if [ -L "${localConfFilePath}" ]
-        then
-          rm ${localConfFilePath}
-        elif [ -f "${localConfFilePath}" ]
-        then
-          mv ${localConfFilePath} ${localConfFilePath}.bak
-        fi
-        ln -s ${localConfFile} ${localConfFilePath}
-      '';
+      preStart =
+        ''
+          mkdir -p /var/lib/zerotier-one/networks.d
+          chmod 700 /var/lib/zerotier-one
+          chown -R root:root /var/lib/zerotier-one
+
+          # cleans up old symlinks also if we unset localConf
+          if [[ -L "${localConfFilePath}" && "$(readlink "${localConfFilePath}")" =~ ^${builtins.storeDir}.* ]]; then
+            rm ${localConfFilePath}
+          fi
+        ''
+        + (concatMapStrings (netId: ''
+          touch "/var/lib/zerotier-one/networks.d/${netId}.conf"
+        '') cfg.joinNetworks)
+        + lib.optionalString (cfg.localConf != { }) ''
+          # in case the user has applied manual changes to the local.conf, we backup the file
+          if [ -f "${localConfFilePath}" ]; then
+            mv ${localConfFilePath} ${localConfFilePath}.bak
+          fi
+          ln -s ${localConfFile} ${localConfFilePath}
+        '';
 
       serviceConfig = {
         ExecStart = "${cfg.package}/bin/zerotier-one -p${toString cfg.port}";
